Séparation de clickedPoint en fetch et map
